+++ src/app/(main)/admin/_components/loading.tsx
@@ -1,0 +1,9 @@
+import { LoadingDots } from '@/components/ui/loading-dots'
+
+export const Loading = () => {
+  return (
+    <div className="flex w-full h-full items-center justify-center">
+      <LoadingDots />
+    </div>
+  )
+}
