+++ src/app/(main)/admin/_components/semester/semester-manage.tsx
@@ -3,10 +3,10 @@ import { useEffect } from 'react'
 import { useSemesterFetch } from '@/services/fetchSemesters'
 import { useAdminSemesterStore } from '@/store/admin-semester'
 
-import { parseSemester } from '../_utils/parseSemester'
-import { ItemLayout } from './item-layout'
-import { AddDialog } from './semester/add-dialog'
-import { ActivityDialog } from './activity/activity-dialog'
+import { parseSemester } from '../../_utils/parseSemester'
+import { ItemLayout } from '../ui/item/item-layout'
+import { AddDialog } from './add-dialog'
+import { ActivityDialog } from '../activity/activity-dialog'
 
 export const SemesterManage = () => {
   const { fetchSemester } = useSemesterFetch()
